[assembly] Add missing dep on blobfs image to flash.sh
In the refactor to using assemble_system(), the //:images build_api
target lost it's transitive dep on the blobfs image.  When the
temporary copy action was added, uefi_disk picked it up, but that's
not present on all builds, so flash.sh also needs a dep on it.

Bug: 82274
Bug: 195576132

Change-Id: I93e237d05962cd8b273f27a53d55fb6a7d0ddaff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565861
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>