[botanist][syslog] Re-add package repo if ssh disconnects.
If we lose SSH connection to the device, but are able to reconnect,
there's a possibility that the disruption was caused by a reboot which
could have removed the package repo config. This would show itself if
the syslog we start streaming in botanist gets interrupted but is able
to restart again (after the reboot). Thus, we can listen for errors from
the syslog streaming and re-add the package repo config so that it will
work for future tests in the shard.

Bug: 81706
Change-Id: I3ce495fc034c43e9b6b473df775071fc1a5009ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562301
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>