[realm-builder] Expose fuchsia.component.Binder from root
This CL changes the framework_intermediary library to expose
the protocol `fuchsia.component.Binder` from the root of a
constructed realm. This change was made to allow users to
start the realm in the event that have no other mechanism to
do so.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400

Change-Id: I7706396c529079ac9448a32ac64a80a722050ed8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564562
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>