[netstack] Fix a flake in DHCP inspect tests
These tests send a packet to the client and then verify that various
stats have been incremented. When the sent packet arrives before the
DHCP client has begun listening for it, the packet is dropped and we
fail to observed the expected stats. Eliminate this race condition by
waiting for a message from the client before sending our packet in the
test runner, which ensures that the client will be ready by the time it
arrives.

Bug: 80569
Multiply: netstack-inspect-integration-test

Change-Id: I60c14140e5d401cbeb21fed423b9bdf42ed13295
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564702
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>