[cm][routing_tests] Test routing of one event at a time
Some of the routing tests for Event capabilities include
a single ExpectedResult for multiple EventSubscriptions.
This is a bit inconvenient when implementing the
RoutingTestModel trait for the static analyzer, so this
change slightly refactors RoutingTest's check_use()
method for events in order to test one event at a time.

Bug: 77649

Change-Id: Ibba2cd2c203a740ca00328d58252bc64b192ca9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557811
Commit-Queue: Laura Peskin <pesk@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>