[starnix] Resolve race in FsNode::create_child
Previously, we could crash in FsNode::create_child if another thread was
dropping the child at the same time as we were trying to fetch it. The
other thread could have entered Drop::drop (having removed the last
refcount) and be waiting on the RwLock. The first thread would then see
a populated OnceCell with a Weak that failed to upgrade. Prior to this
CL, that upgrade failure would result in a panic.

This CL removes the OnceCell and we now use the Weak as the source of
truth for whether the entry in the cache is populated populated.

Change-Id: I245d2bf7f92b6206dd56cfb952ca78ca75089e4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566568
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>