[assembly] Remove vbmeta salt parameter
The transitional salt parameter in the VBMetaConfig is not longer used.
The salt parameter was used for sharing the same salt between two sets
of images, so that they could be binary equivalents. Now that we are not
comparing two sets of images, this is unused and can be removed.

Bug: 77114
Change-Id: Ie5ac5522c32497b31b92982fe54718025410a30d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564324
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>