[fidl][rust] Check vector, string, and table counts
This CL adds a decode_vector_header function to centralize checks for
vectors, strings, and tables:

* To satisfy RFC-0059, it checks that the count is less than 2^32.

* To fix fxbug.dev/82159 (ASan reporting extremely large allocations),
  it fails if the count is larger than the overall message buffer.
  This likely also fixes fxbug.dev/82152.

Since the second check is meant to avoid huge allocations, we prefer a
simple comparison to the message size. There is no benefit to making it
more clever by subtracting the current offset, for example, because
that is only accurate if the vector is the last object in the message.

This also adds more GIDL tests to exercise the various cases for
vectors, strings, and tables.

Bug: 82167 (check vector sizes)
Bug: 82159 (ASan excessive allocation size)
Bug: 82152 (invalid slice from addition overflow)

Change-Id: I26759a31c0d8a9138c70af5839020de00f468fd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565982
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>