[archivist] Reduce log verbosity when handling an error directory ready event
Since we introduced the inspect shard, many test root CMLs had to transitively
include the shard. This led to some log spam which confuses people when
their test is failing for unrelated reasons.

An alternative to this implementation would be to not emit Error events
when a component "exposes" a directory to framework but never actually
serves it. I believe that the current (and tested) implementation in component
manager of emitting Error payloads when the component didn't serve the
directory but exposed it is correct and instead we can just reduce the
log verbosity here.

Fixed: 82010

Change-Id: Idfbfb944e3f2701a166f2feceb5a8b822ffd8f55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566809
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>