[hlcpp] V2 wire format encoding for HLCPP
V2 decoding support was previously added for HLCPP, but not encoding
support. This CL adds encoding support. There currently isn't a
validator that supports v2 validation so this CL runs the v1 validator
and disables tests that break due to differences (only 8).

It is not enabled in the live path and is only enabled for GIDL tests.

Bug: 81995

Change-Id: I01a69a7d2999b4ca6d3f7924ec239878e0b8c868
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564061
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Ian McKellar <ianloic@google.com>