[kernel][x86] page_tables: Test for PTE global bit when queueing inval
When queueing invalidations set the global bit based on a PTE's global
bit, rather than on whether it is a kernel (vs user) address. The two
predicates come out the same way currently, but if there are uses of the
global bit for non-kernel addresses in the future, the current predicate
would be wrong.

Resolves an old TODO from teisenbe@.

Change-Id: Ic745d58a3434c779253a41b1b9c2c4fedeacb915
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566567
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>