[settings] Wait for camera devices to init watcher
The initialization order is not guaranteed between the camera api and
the camera watcher agent, so there may be a valid empty response to
watching the devices. This change introduces a timeout during which the
agent will continue to watch for devices before deciding it can't find a
camera device.

Fixed: 79768
Test: fx test setui_service_tests

Change-Id: If1dd5b1a7792a8fcec30c29e7bb77722262930eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564139
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>