[fidlc] Always specify naming context
Previously, we would only specify the naming context in the IR for
anonymous layouts (since otherwise, it is implicitly just the
naming of the layout). This changes naming context to always
be present, renaming it from maybe_naming_context to naming_context.

This is mainly for clarity, so that things are explicit in the IR
and fidlgen does not need to make the assumption listed above.

Change-Id: I13551930410355d5e639eb0a377777ff5270af68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567301
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>