[zxio] Add zxio_create..() variant to wrap known representation
Some users of the zxio library wish to create a new zxio object with
knowledge of the concrete type and type-specific objects on hand from
sources outside zxio-mediated protocol negotiation. These callers need a
way to inflate a zxio object around the raw pieces directly. The
inception header provided a way to do this by exposing the zxio type's
internal structure and letting callers directly initialize the object
and then use it. This change adds a pure C entry point that exposes the
initialization types but hides the structure inside the zxio library.

As the types needed to initialize an object vary by type, exposing this
operation requires either a large number of entry points or a dynamic
solution. This change uses a type enum and C's variadic arguments
feature to have a dynamically typed function. This also provides a
header-only C++ library that provides type safe wrappers around the
entry point. We should do the same for Rust and other languages wishing
to bind to zxio to make the library more usable.

Only the subset of types that are needed by fdio today are supported.
More types can be added as needed.

Several callers in fdio are updated to use this new entry point instead
of the inception.h provided facilities.

Bug: 43267

Change-Id: I2b88c4952b525f090ce492ed2f73591c3af83131
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549883
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>