[fxfs] Periodically flush files
Adds a per-volume background flushing task which periodically sweeps
through all files and attempts to flush them.

Attempting to flush every file is slightly inefficient, since we always
take a transaction lock when we attempt flushing. We can make this
better by storing a dirty bit in the CachingObjectHandle (or in the
WritebackCache) so no lock needs to be taken on the object, but this can
be done in a follow-up.

Tested: Manually verified that the flush task periodically runs.
Tested: Unit.
Change-Id: Ibad4b4215a729ec3de67630553fae43015e70820
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565807
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>