[fidl][cpp] NaturalClientBase and unified client wrappers
This CL adds |fidl::Client| and |fidl::SharedClient|, the unified
counterparts to |fidl::WireClient| and |fidl::WireSharedClient|. The
interface documentations are largely identical with the exception of the
`.wire()` method and some mentions of wire messages. No natural
messaging layer is generated yet; as a starting point we are testing
that the client wrappers themselves work.

Given a unified client:

  fidl::Client<SomeProtocol> client;

the wire methods are exposed under:

  client.wire()->Foo(...);

the natural methods are exposed under:

  client->Foo(...);

See the design doc linked in fxbug.dev/60240 for more details.

In order to expose two different set of methods, inheritance cannot be
used. We introduce |NaturalClientBase| as the base class for all future
natural client implementations, to be exposed by the `->` operator.
|NaturalClientBase| contains a |NaturalClientMessenger| which borrows
into the |ClientBase| that is part of the wire client implementation
object.

In the future, we could refactor WireClientImpl to also use a messenger
object, instead of inheriting from |ClientBase| and leaking a bunch of
internal methods to the user (fxbug.dev/82085).

Bug: 60240
Test: fx test //src/lib/fidl/cpp
Change-Id: Ia526f839e47a58bba3f99663714ff44632b0e5bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565681
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>