[component_manager] increase timeout for test
The model::component::tests::stop_component_well_behaved_component_stop
test is flaking in CI. It flakes because it takes too long to stop the
component.

The timeouts for this are very short, 5 milliseconds for the stop
timeout, and are likely insufficient when the test bot is under
significant load. To resolve the flake, 10x the timeouts.

Fixed: 82609
Multiply: component-manager-tests
Change-Id: Ibe64730549285f4d28b1606aee4458e8cdd39704
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568121
Fuchsia-Auto-Submit: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>