[fidl][llcpp] Support scoped anonymous layout names
This CL updates LLCPP to generate aliases used to refer
to an anonymous layout from within the layout where it's
defined. For example, in

  type Parent = struct { child struct {}}; };

The inner anonymous layout can be referred to both as
`Child` as well as `Parent::Child`. This is useful since
the scoped name can sometimes be shorter/more readable,
e.g. for response result types or cases when a flattened
name is specified manually (see fxbug.dev/81550).

Fixed: 81471
Test: fx test --host -vo src/lib/fidl/llcpp/tests/wire_types
Test: fx test --host -vo fidlgen_cpp_ir_test
Change-Id: Id672761917db8ac27fd65d71a7e945674b040468
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562604
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Felix Zhu <fcz@google.com>