[archivist] Refactor pipeline initialization for simplification
The archivist initialization code is a bit large and can be logically
subdivided. This tackles the pipeline initialization.

Additionally, this simplifies the work needed to create a new pipeline.
For example, if we tackle fxbug.dev/65801 and create a new pipeline for
sampler metrics, it now takes only a few lines of code to do so while
getting all the instrumentation and boilerplate for free.

Tested: archivist unit tests and integration tests

Change-Id: I2c425edffc1f1495b046a92be0bdd65965fcc810
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567530
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>