[scrutiny] Enable configurable plugin loading.
Prior to this changes plugins were not actually configurable through the
config file. This change allows plugins to register but not load for
certain use cases. This allows more fine grain control in apps as to which
specific plugins are running. It is up to the user of the API to correctly
set the plugins they need and all dependencies. This now allows for the
toolkit plugin to not use the core plugin at all.

Bug: 81088
Change-Id: I5a74082ad3f2f33e01161320b3e75092f81fe603
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565897
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>