[build] add support for "bootstrap" flash target
Bootstrapping a device is the process of taking it from an unknown state
to the point where it can flash Fuchsia using the normal flow.

This doesn't need to be done often - generally only when provisioning a
board in the first place or recovering a bricked device - but it will be
useful to have a standard flash target to support this flow.

An example of a file that may be included here is the GPT, which isn't
expected to change during the normal device lifecycle but does need to
be initially flashed once.

Bug: 64549
Bug: b/188193322
Test: ffx target flash <flash.json> bootstrap
Change-Id: Ibd4f5ba532f69a8e4f923b066a22873f6c50ce9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567447
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>