[benchmarks] Introduce runTestComponent() helper function
Introduce the runTestComponent() helper function and change a test to
use it.

This adds steps to generate summarised versions of fuchsiaperf files
and remove temporary files.  The former might change the results from
this test, because it means we'll now be properly discarding the times
from the initial warmup runs of the test cases (an issue I missed in
https://fuchsia-review.googlesource.com/c/fuchsia/+/559687 and which
is described in src/tests/end_to_end/perf/test/summarize.dart).

This function is similar to runTestCommand() except that it invokes a
component rather than an arbitrary shell command.

Bug: 82465
Test: fx set terminal.x64
  + fx test --e2e network_device_microbenchmarks_test
Change-Id: I85c84d23ff5e8673dfdda09521e1ec050ed85a37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568576
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>