[power] Replace CobaltMetrics with PlatformMetrics
Replaces the existing CobaltMetrics singleton with a more flexible and
useful node called PlatformMetrics.

The primary motivation for this transition is to support capturing the
same historical CPU temperature metrics on Nelson that we are on Astro.
There wasn't previously a clean way to do that considering the node that
publishes this metric on Astro (ThermalPolicy) is not used on Nelson.

This change restructures the CobaltMetrics singleton as a node and in
doing so, configures it to take a more hands-on approach to gather the
required metrics data. Instead of having CPU temperature pushed into it,
the PlatformMetrics node will poll the CPU temperature handler at a
configured interval (1s). This approach is easier to reason about and
eliminates the trap that we'd fallen into previously where we enable the
ThermalPolicy node for the sole reason of capturing these metrics, while
adding hacks to disable the actual ThermalPolicy behavior.

Test: fx power-manager-tests
Bug: b/178334881
Change-Id: Id7d9c4f0e7d9b70d743e0c9e062a4a3068cef5cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563144
Reviewed-by: Onath Claridge <claridge@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>