[assembly] Use assemble_system in zedboot in parallel
We are using the assemble_system template for constructing zedboot in
parallel, and ensuring the resulting ZBI is the same as the current
assembled ZBI.

We are not checking the VBMeta's equivalence, because I got excited and
already yoinked the code that supports injecting a salt
(fxrev.dev/564324). The VBMeta is nothing special for Zedboot, so I
don't think we should be concerned with not checking for its
equivalence.

Bug: 82474
Change-Id: Ie455b3508666b651e39c2f5facf2606032319f4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564848
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>