[driver_manager] Move eager fallback to driver.cc
There are boot parameters that will mark a driver a fallback
driver as a normal driver. Now that a driver's fallback status
is set in driver.cc, this is also where the eager check
should be handled.

This means that eager/fallback checking happens in the same
way for drivers found through the DriverIndex or through
DriverManager.

TEST: fx test ddk-fallback-test
Change-Id: Ibc9346f376bd4ff12a7e5afcefb49ac760dea951
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568181
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>