[sl4f][mic_driver] dart client implementation and integration test for setui_facade.setMicMute
dart client accesses setui_facade.setMicMute through an SL4F request. Test ensures that no errors are thrown when SL4F setMicMute is spoken to. From an in-tree Fuchsia point of view, the test is end-to-end, but strictly speaking it is an integration test.

TODO b/196844246

Change-Id: I9d0039d2a6cf1159b5dacd2004bcc41d7e819484
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567801
Commit-Queue: Isabelle Goralsky <igoralsky@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Cole Alban <colealban@google.com>