[gn-sdk] Enabling auto indexing of package symbols
This adds a global flag `fuchsia_auto_index_symbols`.
When true, the ids.txt file generated from a `fuchsia_package`
is added to the symbolizer index using the `symbol-index` SDK tool.
The Fuchsia symbol index used to symbolize logs and to find source
when using zxdb.

The default value is false, in order to be backwards compatibility
with existing builds, and only makes sense to set to true when building
as a developer that will need to symbolize logs or use the debugger with
locally build code.

To enable it for a build add `fuchsia_auto_index_symbols=true` to
the gn gen args.

Bugs: b/195019412
Change-Id: I459a152703c8808c4c4a8bdd5a831a33af3b2c5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567504
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>