[schema id] update schema id of cmx_schema.json
In Cl 530463 I'd made a typo when changing the id for
src/sys/lib/cm_json/cmx_schema.json by omitting the src
directory, which is part of the path to the file in git.
This CL adds that "src/" part of the path.

Change-Id: Icc8bea512546bef8b116e41bc8e140bf7dcca3ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568987
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>