[kernel][crypto] Make OWNERS noparent and add a couple reviewers
Random number generation is notoriously difficult to test.  As a
precaution, we'd like Security reviewers to give extra scrutiny to
changes that could cause regressions in how the PRNG is set up.

To ensure that we're not creating a single-point-of-failure, we add a
few additional folks from the security team to the list of approvers.

Fixed: 82811
Change-Id: I7f516ea899d4aa61668a43826c4d2ad0030abdec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569359
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>