[lsc] Migrate BindChild to OpenExposedDir
This CL is part of a migration of usages of
fuchsia.sys2.Realm/BindChild to fuchsia.sys2.Realm/OpenExposedDir.
The former has been deprecated.

For this CL in particular, the log-stats integration test
has been changed to explicitly bind to the
`fuchsia.component.Binder` so as to start the `log-stats`
component.

Change-Id: I4f59e3c5a97604f5896d25f0a11c1fd955b529bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563930
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>