[kernel][hypervisor][arm64] Route SError exceptions to EL2
On ARM64, asynchronous SError (system error) exceptions are triggered
external to the CPU, such as a memory access which has passed all the
MMU checks but encounters an error on the memory bus.

The Zircon kernel currently masks SErrors (https://fxbug.dev/72659,
https://fxbug.dev/75834), but Linux guest kernels keep them unmasked.
The result is that if host Zircon encounters an SError exception, it
will remain pending until Zircon finally schedules a Linux guest, which
will spuriously take the exception.

This CL updates the hypervisor's EL2 config to route SErrors taken by
guests to EL2. These are sent to the host Zircon kernel which prints an
error message to the kernel log, and then resumes the guest.

Bug: 57651
Fixed: 75609

Change-Id: Id2f7b9bb8791892fdb63cf7c7cdedc3c53988c9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/565622
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>