[inspect][rust] Add counter for failed allocations
When we run out of memory and cannot add a new page with new blocks we
just return an error. In order to export how often this happens, we're
adding a counter that reports how many failed allocations we've had
since the heap was created.

Bug: 75729
Change-Id: I34e14971d779b317efb0c4b34d8ba280a6db7ef2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569421
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Adrian Hausner <adrianhausner@google.com>