[brcmfmac] Remove unsupported countries from list.
Bug: 82462

As per the HW team and vendor, the following countries
are not supported yet for Sherlock:

KR, TW, IN, SG, MX, NZ

To be in sync with the clm.blob these countries need
to be removed from the list of supported countries.

Test: Existing tests cover the case of setting an
unsupported country code.

Change-Id: I1e268e12fdf4ac41335cc4bde2a89c88c7e07b4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569967
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Tim Kilbourn <tkilbourn@google.com>
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>