[ffx][repository] Work around `fx serve-remote` port collision
When using `fx serve-remote`, when both a local laptop and the remote
workstation may be configured to serve repositories. When this happens,
both repository servers may try use the same port. This currently will
report a warning, but otherwise not give any indication that the tunnel
failed to be created.

To work around this, this patch:

* Changes `ffx repository` to allow the port to be disabled.
* Changes `fx serve-remote` to detect the port collision, and error out.
* Changes `fx add-update-source` to report an error if the server is
  disabled.

Bug: 82788

Change-Id: I7a61b92e83bb2c5531814c0542b5173509113959
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569743
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: James Tucker <raggi@google.com>