[fidl][c] Change decoded representation for v2 unknown envelopes
The v2 wire format introduces an inline envelope representation in
addition to the existing out-of-line representation. As with the
v1 wire format, the decoded form of out of line envelopes are pointers.
Supposing that the decoded form of the inline envelopes were the same
as the wire format, the inline flags bit would overlap with the pointer
data making it impossible to discern whether the envelope is inline
or out of line. By shifting fields around in the decoded representation,
it is possible to have the inline flags bit overlap with the lowest bit
of the pointer, which is always zero due to alignment.

In order to implement HLCPP v2 decoding support, it is important to
have inline and out of line envelopes differentiated.

Bug: 81995

Change-Id: I184fe909dda3f50015c032e4f9bbffb79a847d36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569302
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>