[camera][sensor] Fix num_modes_ initialization
num_modes_ is not properly initialized to any reasonable
value, but is used in a critical range check.

Since num_modes_ is only used in one place, and that size is
easily fetched from the said array, we choose to eliminate
num_modes_ completely.

Bug: b/196701267
Test: Force an out-of-range value to see if the range check fails.
      Force an in-range value to see if the range check succeeds.
Change-Id: Ib16180547dff7f55bc7e9e3f54257a3a6795b305
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569121
Commit-Queue: Ernest Hua <ernesthua@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>