[recovery] Use rive instead of shed
In preparation for removing shed.

If an out-of-tree client has specified a shed file, this
implementation will work but will not draw a logo.

While there, fixed the fkey emulation of media keys that
was broken by the implementation of key repeat events
and improved the centering of the countdown numbers
a little bit.

Testing: manual
Change-Id: Ib9e4852c143cd8536ca1ff5ddab905797ae274c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567382
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>