[realm_builder] improve realm destruction
Adds a drop implementation for `RealmInstance` that will spawn a new
async task that holds ownership of the mocks runner while it waits for
realm destruction to finish, such that mocks are stopped in an orderly
fashion instead of all at once.

Also adds a `destroy` function to `RealmInstance`, which can be called
to block until a realm is successfully torn down. Much like the new drop
implementation, this function takes care to keep the mocks runner task
alive long enough for the mocks to stop in an orderly fashion.

Fixed: 78987
Change-Id: I21c1481bd4501a4d052211f580fe4b86353b8b30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563576
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Gary Bressler <geb@google.com>