[hidctl] Ensure worker thread has finished before finishing destructor.
When HidDevice is being destroyed, we don't currently wait for the
worker thread to fully finish before completing the destructor. This
leads to races where the still-running worker thread attempts to access
state in the now-deleted HidDevice instance.

This doesn't appear to affect the driver proper (as the DDK will call
`DdkRelease`, which _does_ wait for the thread), but does trigger in
unit tests.

Update the class to ensure we join on the thread in the destructor.

Fixed: 80427
Change-Id: I8bf08c2f7950f0af9c7ab794c125694741c7acfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570321
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>